# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "E-commerce API Documentation",
        "description": "Swagger Docs for Dashboard APIs (Admin Panel)",
        "contact": {
            "name": "API Support",
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Local server"
        }
    ],
    "paths": {
...
        "/e_commerce/banner/{id}": {
            "get": {
                "tags": [
                    "E-Commerce - Home"
                ],
                "summary": "Get product associated with a specific banner",
                "operationId": "d7fbb10e4e8d1a9a71d21751e24749d6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Banner ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 3
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product retrieved successfully"
                    },
                    "404": {
                        "description": "Banner or product not found"
                    }
                }
            }
        },
        "/e_commerce/product/{id}": {
            "get": {
                "tags": [
                    "E-Commerce - Home"
                ],
                "summary": "Get product details by ID",
                "operationId": "5db0595ceb8cf82f9620699388311028",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 5
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product retrieved successfully"
                    },
                    "404": {
                        "description": "Product not found"
                    }
                }
            }
        },
        "/e_commerce/category/{id}": {
            "get": {
                "tags": [
                    "E-Commerce - Home"
                ],
                "summary": "Get products of a specific category",
                "operationId": "885c8a8d9a30fa936643e3f83a7647dc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Category ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 2
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Products in category retrieved"
                    },
                    "404": {
                        "description": "Category or products not found"
                    }
                }
            }
        },
...
    },
...
}